% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbills.R
\name{getBills}
\alias{getBills}
\title{Gets the bills depending on congress, branch, and offsets
key is the api key
congress it the number of the congress
branch is the House or the Senate branch
numFrom is the offset number from
numTo is the offset number to}
\usage{
getBills(key, congress, branch, type, numFrom, numTo)
}
\description{
Gets the bills depending on congress, branch, and offsets
key is the api key
congress it the number of the congress
branch is the House or the Senate branch
numFrom is the offset number from
numTo is the offset number to
}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/}
}
\author{
Nadine Fischoff, \email{nadine.khattak@gmail.com}
}
\keyword{api,}
\keyword{bills}
\keyword{congress,}
\keyword{propublica,}
